Commit title: add new functional
Commit description: 3 classes have been added:

1. Task
2. TaskSettingsDialog
3. TimerDialog
Task is the main class that is a descendant of QWidget. It has a button to start a timer,
which, when pressed, calls the TimerDialog class, the name of the task, when clicked,
the TaskSettingsDialog class is called, as well as a label with star icons (temporarily)
indicating the duration tasks. Also, specifically for this class, a new Layout was added to listPage,
which has its own ScrollArea.

TaskSettingsDialog - inherits from QDialog, has 2 fields and 2 buttons.
The fields change the name and duration of the attached Task object. The first button confirms changes,
the second button deletes the Task objectю

TimerDialog - also inherits from QDialog. At the moment it is a dummy that needs to be completed