Commit title: reworked dialogs
Commit description: The dialog boxes have been redesigned and breakTimer now works.
The TimerDialog interface has been temporarily redesigned, and a timer start button
has also been added to the TaskSettingsDialog. workDuration and breakDuration were
added to the Task class because It is planned to change them inside the object
itself so that the are remembered.

Issues: It still barely works and everything needs to be completed. Duration
variables in Task are now dummies. We need to completely reconsider the system of
durations themselves, this is necessary